backend/requirements.txt
# Clarity by Plinng - Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0

# Data Validation
pydantic>=2.5.0

# Environment Variables
python-dotenv>=1.0.0

# AI / Gemini API
google-genai>=1.0.0

# Web Scraping (for Research Agent)
httpx>=0.27.0
beautifulsoup4>=4.12.0

# LangChain & LangSmith
langchain>=0.3.0
langchain-google-genai>=2.0.0
langsmith>=0.2.0
